refactor: move and adjust recent ideas fetch function
Move fetch function from recent ideas to helpers so it can be reused
by the Ideas block. Also moves datastore to helpers.

Change recent ideas fetch function to return the fragment instead of
appending it.
