+++ blocks/recent-ideas/recent-ideas.js
@@ -1,69 +1,4 @@
-import { createOptimizedPicture } from "../../scripts/aem.js";
-import { buildCard } from "../card/card.js";
-import { dataStore } from "../../scripts/dataStore.js";
-
-/**
- * The settings object used for fetching and creating the list of ideas.
- * @typedef {object} RecentIdeasSettings
- * @property {string} tagName - Fetch articles with this tag, or "All" to fetch all articles.
- * @property {number} maxArticles - Max articles to fetch or -1 for infinite.
- * @property {string} gridItemClass - Class for each grid item that determines layout; e.g. "grid-item--25" for four-up layout.
- * @property {boolean} hasHorizontalScroll - Has horizontal scroll at mobile.
- */
-
-/**
- * Fetch article data, create card markup, and append it.
- * @param {HTMLElement} cardsParent Parent element to append cards to.
- * @param {RecentIdeasSettings} settings Settings for what is fetched and how many are fetched.
- */
-const fetchDataAndBuildCards = async (cardsParent, settings) => {
-    // Contains all of the grid items and cards to be appended at the end.
-    const articleHolder = document.createDocumentFragment();
-
-    try {
-        // Fetch articles from JSON.
-        // The articles should already be sorted with the latest (published date) first in the returned data.
-        const articles = await dataStore.getData(dataStore.commonEndpoints.ideas);
-        let filteredArticles = articles.data;
-
-        // Filter by the specific tag.
-        const tagToFind = settings.tagName.trim().toLowerCase();
-        if (tagToFind !== 'all') {
-            filteredArticles = filteredArticles.filter(item => item.tag.trim().toLowerCase() === tagToFind);
-        }
-
-        // Sort by published date (serial number or timestamp), with the latest dates first.
-        filteredArticles = filteredArticles.sort((a, b) => b.publishedDate - parseInt(a.publishedDate, 10));
-
-        // Limit the number of results displayed.
-        if (settings.maxArticles !== -1) {
-            filteredArticles = filteredArticles.slice(0, Number.isInteger(settings.maxArticles) ? settings.maxArticles : 4);
-        }
-
-        // Build markup.
-        filteredArticles.forEach(article => {
-
-            // Create card and append.
-            const articleImageUrl = article.image.trim();
-            const card = buildCard({
-                img: articleImageUrl ? createOptimizedPicture(articleImageUrl) : '',
-                textContent: [
-                    article.title,
-                    article.description
-                ],
-                url: article.path.trim(),
-            });
-            card.classList.add('grid-item', settings.gridItemClass);
-            articleHolder.append(card);
-        });
-        // Append everything within our fragment to this parent.
-        cardsParent.append(articleHolder);
-    } catch (error) {
-        // eslint-disable-next-line no-console
-        console.error(`Error fetching and displaying the articles: ${error}`)
-        return;
-    }
-};
+import { fetchAndBuildIdeas } from "../../scripts/helpers/fetchAndBuildIdeas.js";
 
 /*
  * Recent Ideas Block
@@ -91,5 +26,9 @@ export default function decorate(block) {
     block.parentElement.replaceWith(newBlock);
 
     // Fetch and add markup for articles (async).
-    fetchDataAndBuildCards(newBlock, settings);
+    const fetchAndAppend = async () => {
+        const fragment = await fetchAndBuildIdeas(settings);
+        newBlock.append(fragment);
+    };
+    fetchAndAppend();
 }
